Revamp api and functions for stages

+++ src/app/api/stages/delete/route.ts
@@ -1,17 +1,23 @@
 import prisma from "@/db/prisma";
-import { NextRequest, NextResponse } from "next/server";
+import { NextRequest } from "next/server";
+import { API_RESPONSE_MESSAGES } from "../../(lib)/consts";
+import { auth } from "@/app/auth";
 
 export const DELETE = async (req: NextRequest) => {
-  const { id } = await req.json();
+  const session = await auth();
+  if (!session)
+    return Response.json({ message: API_RESPONSE_MESSAGES.ERROR.UNAUTHORIZED }, { status: 500 });
+
+  const { userId, id } = await req.json();
 
   if (!id) {
-    return NextResponse.json({ message: "Stage ID cannot be empty" }, { status: 500 });
+    return Response.json({ message: API_RESPONSE_MESSAGES.ERROR.BAD_REQUEST }, { status: 500 });
   }
 
   // Also delete the constrain
   await prisma.jobProgress.deleteMany({
     where: { stageId: id },
   });
-  const data = await prisma.stage.delete({ where: { id } });
-  return NextResponse.json(data, { status: 200 });
+  await prisma.stage.delete({ where: { userId, id } });
+  return Response.json({ success: true }, { status: 200 });
 };
